src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-core
azure-identity
azure-storage-queue
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.0
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
multidict==6.0.4
packaging==24.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv
requests
six==1.16.0
typing_extensions==4.6.2
urllib3
yarl==1.9.2
